requirements.txt
instagrapi==2.1.2
python-dotenv==1.0.0
requests==2.31.0
dateparser==1.2.0
price-parser==0.3.4
Pillow==10.2.0
